Fix the undefined behaviour that was actually undefined
Stages 1 and 2 of docs/undefined-behaviour.md. UBSan findings in a
startup run: 138 before, 108 after, and everything removed was a real
defect rather than a suppression.

The two that mattered most are parsers of untrusted data. The Wolf and
ROTT texture sniffers decide what a lump is by multiplying two 16-bit
values taken straight from its first four bytes:

    if(file.GetLength() == Width*Height+4)   // 64768 * 47104

Both promote to int and the product does not fit in one. This overflows
on Corridor 7's own archive, never mind a resource pack a player
downloaded from somebody else, which E13 made possible. Both now compute
in a type that holds the result.

The scanner copied an uninitialized bool -- observed holding 3 -- on
every DECORATE parse, which is undefined and which a compiler may
"know" cannot happen. ParserState now initialises its members.

Four negative left shifts became multiplications: audio samples in
id_sd and file_audiomus are signed and routinely negative, and
r_sprites shifts a view offset that goes negative whenever the view is
wider than it is centred. dbopl's guessAdd * mul overflowed at
591053 * 4158 and is widened. memset and memcpy through null pointers
in w_wad, and three qsorts that took the address of element zero of an
empty array, are guarded.

Both alignment faults were one allocation carved into differently typed
regions with no padding between them. GameMap::SetupLinks laid a bool
array, a short array and a pointer table in one block computed from
sizes alone, so an odd zone count put eight-byte pointers on an odd
address. FRemapTable::Alloc put the PalEntry palette at exactly count
bytes, so every font whose colour count is not a multiple of four --
which is most of them -- read and wrote four-byte entries misaligned;
that one allocation was the source of six findings in v_font and
v_palette. Both now round each region up to what the next type needs.

tools/test_untrusted_lumps.sh builds a pk3 holding a lump that claims to
be 64768 by 47104, loads it under the sanitizers and requires that no
sniffer overflows on it. The survey log from before these fixes shows
those same two sniffers overflowing on stock data, so the check has
something to catch.

What remains is 108 findings in two idioms and nothing else: 91
downcasts through ZDoom's defaults object, and 17 offsets computed
through a null or fake pointer. Neither is fixable without replacing
machinery this fork did not write, and both are stage 3's business --
a suppression list that makes any new finding fail the gate rather than
join a crowd of ninety.

Full suite: 105 passed, 0 failed, 4 skipped.

+++ src/textures/wolfrawtexture.cpp
@@ -81,14 +81,23 @@ static bool CheckIfWolfRaw(FileReader &file)
 	file.Seek(0, SEEK_SET);
 	file.Read(header, 4);
 
-	WORD Width = LittleShort(header[0]);
-	WORD Height = LittleShort(header[1]);
-	if(file.GetLength() == Width*Height+4) // Raw page
+	// The product in a type that can hold it.
+	//
+	// Width and Height are read straight out of the first four bytes of the
+	// lump, and both promote to int, so two large 16-bit values multiply to
+	// something an int cannot represent -- 64768 * 47104, in the archive this
+	// game ships with. That is undefined behaviour in a routine whose whole
+	// job is to decide what an untrusted file is, and it runs over every lump
+	// in every archive the engine opens, including resource packs a player
+	// downloaded from somebody else.
+	unsigned int Width = LittleShort(header[0]);
+	unsigned int Height = LittleShort(header[1]);
+	if((uint64_t)file.GetLength() == (uint64_t)Width*Height+4) // Raw page
 		return true;
 
 	Width = BigShort(header[0]);
 	Height = BigShort(header[1]);
-	if(file.GetLength() == Width*Height+4) // Mac raw
+	if((uint64_t)file.GetLength() == (uint64_t)Width*Height+4) // Mac raw
 		return true;
 	return false;
 }
